The implementation has been changed a bit and a new environment variable introduced.
$MINCACHETTL specifies the minimum time (in seconds) we will
cache a RR. However clients will be sent an ever decreasing TTL until
the RR has expired - at which time the server will need to query for
a fresh RR.

$MINTTL specifies the minimum time (in seconds) the client will see. The
server will query for a new RR when it really expires but modifies the
answer (again) for the client.

Those two environment variables can be used together but take care that
the maximum ttl for a RR can be $MINCACHETTL + $MINTTL - 1 if a client
asks for a RR one second before it's supposed to expire.